+++ src/Scriptbook/Public/Scriptbook/Get-WorkflowContext.ps1
@@ -1,0 +1,20 @@
+<#
+.SYNOPSIS
+Retrieves Workflow Context
+
+.DESCRIPTION
+Retrieves Workflow Context
+
+.EXAMPLE
+
+$ctx = Get-WorkflowContext
+
+#>
+function Get-WorkflowContext
+{
+    if (!(Get-Variable -Name Context -ErrorAction Ignore -Scope Global))
+    {
+        Set-Variable -Name Context -Value @{ } -Scope Global
+    }
+    return $Global:Context
+}