+++ src/Scriptbook/Public/Scriptbook/Get-WorkflowSetting.ps1
@@ -1,0 +1,27 @@
+<#
+.SYNOPSIS
+Retrieves one Workflow Setting
+
+.DESCRIPTION
+Retrieves one Workflow Setting like InTest mode
+
+.PARAMETER Name
+Name of Setting
+
+.EXAMPLE
+
+Get-WorkflowSetting -Name 'InTest'
+
+#>
+function Get-WorkflowSetting([ValidateNotNullOrEmpty()]$Name)
+{
+    $ctx = Get-RootContext
+    if ($Name -eq 'InTest')
+    {
+        return $ctx.InTest
+    }
+    else
+    {
+        Throw "Workflow Setting '$Name' not recognized"
+    }
+}