+++ src/Scriptbook/Public/Scriptbook/Out-ScriptbookHost.ps1
@@ -1,3 +1,19 @@
+<#
+.SYNOPSIS
+Writes to the ConsoleHost
+
+.DESCRIPTION
+Writes to the ConsoleHost
+
+.PARAMETER InputObject
+Value to write to Console Host
+
+.EXAMPLE
+
+'Hello' | Out-ScriptbookHost
+
+#>
+
 Set-Alias -Name Out-Info -Value Out-ScriptbookHost -Scope Global -Force -WhatIf:$false
 function Out-ScriptbookHost
 {